AI/requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=3.4.4
opencv-contrib-python==4.8.1.78
Pillow>=8.3.2
pytesseract>=0.3.8
easyocr>=1.6.2
spacy>=3.5.0
google-generativeai>=0.3.0

# Face Recognition
face-recognition>=1.3.0
dlib>=19.22.0

# Deep Learning
torch>=2.0.0
torchvision>=0.16.2
transformers>=4.30.0
sentencepiece>=0.1.99

# Lab Equipment Recognition
google-api-python-client>=2.0.0  # For YouTube API
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6

# Project utilities
pathlib>=1.0.1
python-dotenv>=0.19.2
requests>=2.28.1

# Image processing
scikit-image>=0.19.0
matplotlib>=3.4.3
pandas>=1.5.0

# Development dependencies
pytest>=7.0.0
black>=22.3.0
flake8>=4.0.1
isort>=5.10.1

# Optional but recommended
tqdm>=4.65.0  # For progress bars

# OpenCV GUI dependencies
opencv-python-headless>=4.5.3

# For Linux systems, you might need:
# python-tk
# python3-pil.imagetk

# PDF Processing
pdf2image>=1.16.3
# Note: poppler-utils is a system dependency, not a Python package

# NLP
nltk>=3.6.7

# MediaPipe and Protobuf with compatible versions
mediapipe==0.10.11
protobuf==3.20.3

# Specific versions for compatibility
typing-extensions>=4.6.0
google-api-core==1.34.1
grpcio-status==1.49.0rc1
pydantic>=2.0.0
retina-face>=0.0.17


flask
pymongo 
cloudinary 
python-dotenv

deepface

# -e .